Fixed the way that files are loaded. When maven builds a jar, it puts everything at the root directory. So where we were calling a file by "/res/xxxx/xxx" after building a jar the file would actually be located in "xxxx/xxx"
Instead of loading the file and then loading the BufferedReader, we now
should just be doing this 

BufferedReader = new BufferedReader(new
InputStreamReader(getClass().getResourceAsStream(filePath)));

where the file path is "/xxxx/xxx" not "/res/xxxx/xxx" or whatever. 

I also fixed something else but it was really small and I don't remember
what it was so i can't comment on it.