Added Explosion graphics. Needs to be implemented in a seperate object type or something but it works fine the way that it is.
Added an object "queue" in the handler to avoid concurrent modification.
Basically when you add object, it adds the object to a queued list. The
objects waiting to be added are added when tick is called. This should
have been in here a long time ago i feel. 

Added new method in Animation that allows you to create a new animation
without specifying the specific cells you wish to use, it will just use
the entire sprite sheet. 

--New bugs: After adding the queue for objects some graphics aren't
loaded. Nothing serious though, just in the intro animation and so I
didn't feel it as a priority to fix that stuff.